Move deserialization helper into distributed.protocol
